requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gmft==0.1.0
pandas==2.1.3
Pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 